refactor(grfbrowser): Extract file_tree.go from main.go
Move file tree management to separate file:
- buildFileTree: Create virtual folder structure
- sortTree: Sort directories first, then alphabetically
- matchesFilter: File type filter matching
- matchesSearch: Search pattern with wildcard support
- countFilteredFiles: Count matching files
- rebuildTree: Rebuild after filter changes
- renderFileTree: Render tree view
- renderTreeNode: Recursive tree node rendering

Part of ADR-012 Stage 0: GRF Browser refactoring.
